Ensure a node is open when writing to a buffer
Instead of calling `panic!()` when if someone is using the API wrong,
just make sure the `Node` is opened before calling
`Node::write_to_buffer`. This way the caller don't have to know the
current state of the struct.

This allows us to also make `Node::open` private, remove the C wrapper
functions and remove a FUSE callback.