Only add -R linker flag on Linux
When building the Debian package we expect that the `libarchivefs.so`
library can be found in the `/usr/local/opt/lib/archivefs/lib` path, but
when building it for macOS (usually for development) we want it to use
the relative path in the `target/` path.

This should probably be extracted into some "build for production"
concept instead, but works for now.