Always use archive_read_open_filenames when opening archives
Since opening one file is the same as opening a list of archives with
only one item in it, we can generalize this and use one less API from
`libarchive`.

This was done when trying to hunt down a memory leak that occured in
the `archive_open_and_read_from_path` function, and generalizing it to
always have the same code path makes it easier to debug when running
`valgrind`, since now you can re-produce it without having to use a
multi-part RAR-file.