Fix memory leak in ffi::archive_open_and_read_from_path
When we were using the `archive_read_open_filenames` function, we needed
to pass a `*const *const c_char` pointer to the function. To get this
pointer, we had to do some work to convert all our `String` into `*const
c_char`. When doing this, we lost ownership of the intermediate strings,
causing them to leak as time went on.

This meant that archives that consisted of multiple part-files would
leak faster, since we were leaking the string of the filenames (and more
filers = more file names).

This changes the conversion to ptr in a way that doesn't leak memory.

Closes https://github.com/bzf/archivefs/issues/21